refactor: id primitive type으로 변경 + DTO하나에 글,사진같이 받게 변경
