+++ src/main/java/com/example/flabcaloriecounter/feed/application/port/in/FeedUseCase.java
@@ -1,12 +1,13 @@
 package com.example.flabcaloriecounter.feed.application.port.in;
 
-import java.util.List;
+import java.util.Optional;
 
-import org.springframework.web.multipart.MultipartFile;
-
-import com.example.flabcaloriecounter.feed.application.port.in.response.FeedDto;
+import com.example.flabcaloriecounter.feed.application.port.in.dto.FeedDto;
+import com.example.flabcaloriecounter.feed.domain.Feed;
 
 public interface FeedUseCase {
 
-	void write(final FeedDto feedDto, final List<MultipartFile> photos);
+	void write(final FeedDto feedDto, final long userId);
+
+	Optional<Feed> findByFeedId(final long feedId);
 }
